+++ packages/taro-rn/src/api/utils/index.js
@@ -2,7 +2,7 @@ export function generateUnSupportApi (errText, fnNames) {
   const res = {}
   fnNames.forEach((fnName) => {
     res[fnName] = function () {
-      throw new Error(errText)
+      throw new Error(`${errText} ##  ${JSON.stringify(arguments)}`)
     }
   })
   return res