+++ packages/taro/types/index.d.ts
@@ -415,6 +415,9 @@ declare namespace Taro {
 
   function render(component: Component | JSX.Element, element: Element | null)
 
+  function internal_safe_set (...arg): any
+  function internal_safe_get (...arg): any
+
   function pxTransform(size: number): string
 
   /**
@@ -559,7 +562,7 @@ declare namespace Taro {
     }
   }
   /**
-   * 发起网络请求。**使用前请先阅读[说明](https://developers.weixin.qq.com/miniprogram/dev/api/api-network.html)**。
+   * 发起网络请求。**使用前请先阅读[说明](https://developers.weixin.qq.com/miniprogram/dev/api/network/request/wx.request.html)**。
    *
    * **返回值：**
    *
@@ -610,7 +613,7 @@ declare namespace Taro {
    *
    *     requestTask.abort() // 取消请求任务
    *     ```
-   * @see https://developers.weixin.qq.com/miniprogram/dev/api/network-request.html#wxrequestobject
+   * @see https://developers.weixin.qq.com/miniprogram/dev/api/network/request/wx.request.html
    */
   function request<T = any, U = any>(OBJECT: request.Param<U>): Promise<request.Promised<T>>